SystemCode/Integrations/requirements/requirements1_with_drf.txt
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
pip-autoremove==0.9.1
pyke==1.1.1
pytz==2019.3
sqlparse==0.3.1
